pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexo-logging"
version = "0.0.3"
description = "Logging package for Nexo"
authors = [
    { name = "Agra Bima Yuda", email = "agra@nexmedis.com" }
]
license = { text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "black>=25.1.0",
    "cachetools>=5.5.2",
    "certifi>=2025.8.3",
    "cfgv>=3.4.0",
    "charset-normalizer>=3.4.3",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "distlib>=0.4.0",
    "filelock>=3.19.1",
    "google-api-core>=2.25.1",
    "google-auth>=2.40.3",
    "google-cloud-appengine-logging>=1.6.2",
    "google-cloud-audit-log>=0.3.2",
    "google-cloud-core>=2.4.3",
    "google-cloud-logging>=3.12.1",
    "googleapis-common-protos>=1.70.0",
    "grpc-google-iam-v1>=0.14.2",
    "grpcio>=1.75.0",
    "grpcio-status>=1.75.0",
    "identify>=2.6.14",
    "idna>=3.10",
    "importlib_metadata>=8.7.0",
    "mypy_extensions>=1.1.0",
    "nexo-types>=0.0.2",
    "nodeenv>=1.9.1",
    "opentelemetry-api>=1.37.0",
    "packaging>=25.0",
    "pathspec>=0.12.1",
    "platformdirs>=4.4.0",
    "pre_commit>=4.3.0",
    "proto-plus>=1.26.1",
    "protobuf>=6.32.1",
    "pyasn1>=0.6.1",
    "pyasn1_modules>=0.4.2",
    "pydantic>=2.11.9",
    "pydantic_core>=2.33.2",
    "pytokens>=0.1.10",
    "PyYAML>=6.0.2",
    "requests>=2.32.5",
    "rsa>=4.9.1",
    "typing-inspection>=0.4.1",
    "typing_extensions>=4.15.0",
    "urllib3>=2.5.0",
    "virtualenv>=20.34.0",
    "zipp>=3.23.0",
]

[tool.setuptools]
packages = ["nexo.logging"]

[tool.setuptools.package-data]
"nexo.logging" = ["*.json", "*.yaml"]

[tool.setuptools.package-dir]
"nexo.logging" = "src"
